Split form app
Super basic react-jsonschema-form
