+++ search/src/index.js
@@ -3,14 +3,12 @@ import ReactDOM from 'react-dom';
 import {BrowserRouter as Router, Route, Link} from 'react-router-dom';
 import './index.css';
 import App from './App';
-import Form from './Form';
 import * as serviceWorker from './serviceWorker';
 
 //import RouterMapping from './RouterMapping';
 const RouterMapping = () => (
   <Router>
     <Route exact path='/' component={App} />
-    <Route path='/Form,' component={Form} />
 
   </Router>
 );
